fix: remove inflex dependency (#34)
Remove the dependency as it is not well maintained and has an open
but with OTP 28.  As this dependency is only used in one spot for a simple
pluralization, it is safe to remove without replacement.

Pin the project's Elixir version to the same version in the Dockerfile and CI workflows.

Add the `openssh-client` dependency to the devcontainer
so that local devcontainers can use SSH to interact with the GitHub repo.